Ultimas Alterações

+++ APITracker_Service/Entities/EnderecoAPI.cs
@@ -15,7 +15,7 @@ public class EnderecoAPI
     public string Endereco { get; set; }
     public int StatusCode { get; set; }
     public string Error { get; set; }
-    public int TimeOutEmMinutos { get; set; }
+    public double TimeOutEmMinutos { get; set; }
     public Method Method { get; set; }
     public string Body { get; set; }
     public Ambiente Ambiente { get; set; }
